fix design and path name
