+++ src/components/Login/Login.js
@@ -77,9 +77,9 @@ const Login = () => {
             value={password}
             name="password"
           />
-          <div className="forget-password-link-wrapper">
-            <Link to="/forget-password">
-              <span className="forget-password-link">
+          <div className="reset-password-link-wrapper">
+            <Link to="/reset-password">
+              <span className="reset-password-link">
                 パスワードを忘れた方はこちら
               </span>
             </Link>
