+++ src/components/Login/Login.scss
@@ -76,11 +76,11 @@
       }
     }
 
-    .forget-password-link-wrapper {
+    .reset-password-link-wrapper {
       text-align: right;
       margin-top: -12px;
 
-      .forget-password-link {
+      .reset-password-link {
         font-size: 14px;
         color: $color_primary;
         text-decoration: underline;
