+++ src/components/ResetPassword/ConfirmSentMail.js
