+++ src/components/ResetPassword/ConfirmSentMail.scss
