+++ src/components/ResetPassword/ResetPassword.scss
@@ -1,7 +1,7 @@
 @import "../../sass_config/color.scss";
 @import "../../sass_config/mixin.scss";
 
-.main-ForgetPassword {
+.main-ResetPassword {
   width: 100vw;
   display: flex;
 
@@ -10,14 +10,17 @@
     display: flex;
     align-items: center;
     flex-direction: column;
-    @include tablet {
-      width: 335px;
-    }
 
     .rightBox-container {
       width: 810px;
       margin: 0 auto;
       text-align: center;
+
+      @include tablet {
+        width: 100%;
+        padding: 0 20px;
+        margin: 0;
+      }
 
       .title {
         margin: 80px auto 8px;
@@ -41,7 +44,7 @@
         margin-top: 40px;
         text-align: left;
         @include tablet {
-          width: 335px;
+          width: 100%;
           margin-top: 32px;
         }
 
@@ -60,7 +63,7 @@
           text-align: center;
           color: $warning;
         }
-        .ForgetPassword-buttons {
+        .ResetPassword-buttons {
           display: flex;
           flex-direction: column;
           justify-content: center;
