+++ src/router/AppRouter.js
@@ -21,8 +21,8 @@ import NotFoundPage from "../components/notFoundPage/NotFoundPage";
 import PublicRoute from "./PublicRoute";
 import PrivateRoute from "./PrivateRoute";
 import LogoutRoute from "./LogoutRoute";
-import ForgetPassword from "../components/ForgetPassword/ForgetPassword";
-import ConfirmSentMail from "../components/ForgetPassword/ConfirmSentMail";
+import ForgetPassword from "../components/ResetPassword/ResetPassword";
+import ConfirmSentMail from "../components/ResetPassword/ConfirmSentMail";
 import { isMobile } from "react-device-detect";
 
 const history = createBrowserHistory();
@@ -57,11 +57,8 @@ const AppRouter = () => (
       <PublicRoute path="/joblisting/:jobId" component={JobListing} />
       <PublicRoute path="/contact" component={ContactForm} />
       <LogoutRoute path="/login" component={Login} />
-      <LogoutRoute path="/forget-password" component={ForgetPassword} />
-      <LogoutRoute
-        path="/forget-password-confirm"
-        component={ConfirmSentMail}
-      />
+      <LogoutRoute path="/reset-password" component={ForgetPassword} />
+      <LogoutRoute path="/reset-password-confirm" component={ConfirmSentMail} />
       <PublicRoute path="*" component={NotFoundPage} />
     </Switch>
   </Router>
